Inline associationUpdate into mapUpdateArguments
mapUpdateArguments takes more than just associations as updates after
the pipe and it's more compact and efficient to parse the prefixes
together.
