+++ gen/org/elixir_lang/psi/ElixirMapUpdateArguments.java
@@ -3,11 +3,23 @@ package org.elixir_lang.psi;
 
 import com.ericsson.otp.erlang.OtpErlangObject;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
 
 public interface ElixirMapUpdateArguments extends Quotable {
 
+  @Nullable
+  ElixirAlias getAlias();
+
+  @Nullable
+  ElixirAtom getAtom();
+
   @NotNull
-  ElixirAssociationUpdate getAssociationUpdate();
+  List<ElixirMatchedExpression> getMatchedExpressionList();
+
+  @NotNull
+  ElixirPipeInfixOperator getPipeInfixOperator();
 
   @NotNull
   OtpErlangObject quote();
