--- gen/org/elixir_lang/psi/impl/ElixirAssociationUpdateImpl.java
@@ -1,55 +1,0 @@
-// This is a generated file. Not intended for manual editing.
-package org.elixir_lang.psi.impl;
-
-import com.ericsson.otp.erlang.OtpErlangObject;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
-import com.intellij.lang.ASTNode;
-import com.intellij.psi.PsiElementVisitor;
-import com.intellij.psi.util.PsiTreeUtil;
-import org.elixir_lang.psi.*;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import java.util.List;
-
-public class ElixirAssociationUpdateImpl extends ASTWrapperPsiElement implements ElixirAssociationUpdate {
-
-  public ElixirAssociationUpdateImpl(ASTNode node) {
-    super(node);
-  }
-
-  public void accept(@NotNull PsiElementVisitor visitor) {
-    if (visitor instanceof ElixirVisitor) ((ElixirVisitor)visitor).visitAssociationUpdate(this);
-    else super.accept(visitor);
-  }
-
-  @Override
-  @Nullable
-  public ElixirAlias getAlias() {
-    return findChildByClass(ElixirAlias.class);
-  }
-
-  @Override
-  @Nullable
-  public ElixirAtom getAtom() {
-    return findChildByClass(ElixirAtom.class);
-  }
-
-  @Override
-  @NotNull
-  public List<ElixirMatchedExpression> getMatchedExpressionList() {
-    return PsiTreeUtil.getChildrenOfTypeAsList(this, ElixirMatchedExpression.class);
-  }
-
-  @Override
-  @NotNull
-  public ElixirPipeInfixOperator getPipeInfixOperator() {
-    return findNotNullChildByClass(ElixirPipeInfixOperator.class);
-  }
-
-  @NotNull
-  public OtpErlangObject quote() {
-    return ElixirPsiImplUtil.quote(this);
-  }
-
-}
