+++ testData/org/elixir_lang/parser_definition/map_operation_parsing_test_case/MapUpdateArgumentsWithTrailingComma.txt
@@ -6,21 +6,20 @@ Elixir File(0,28)
         PsiElement(OPENING_CURLY)('{')(1,2)
         PsiWhiteSpace(' ')(2,3)
         ElixirMapUpdateArgumentsImpl(MAP_UPDATE_ARGUMENTS)(3,26)
-          ElixirAssociationUpdateImpl(ASSOCIATION_UPDATE)(3,25)
-            ElixirVariableImpl(VARIABLE)(3,10)
-              PsiElement(identifier)('current')(3,10)
-            PsiWhiteSpace(' ')(10,11)
-            ElixirPipeInfixOperatorImpl(PIPE_INFIX_OPERATOR)(11,12)
-              PsiElement(PIPE_OPERATOR)('|')(11,12)
-            PsiWhiteSpace(' ')(12,13)
-            ElixirMatchedAssociationOperationImpl(MATCHED_ASSOCIATION_OPERATION)(13,25)
-              ElixirVariableImpl(VARIABLE)(13,16)
-                PsiElement(identifier)('key')(13,16)
-              PsiWhiteSpace(' ')(16,17)
-              PsiElement(ASSOCIATION_OPERATOR)('=>')(17,19)
-              PsiWhiteSpace(' ')(19,20)
-              ElixirVariableImpl(VARIABLE)(20,25)
-                PsiElement(identifier)('value')(20,25)
+          ElixirVariableImpl(VARIABLE)(3,10)
+            PsiElement(identifier)('current')(3,10)
+          PsiWhiteSpace(' ')(10,11)
+          ElixirPipeInfixOperatorImpl(PIPE_INFIX_OPERATOR)(11,12)
+            PsiElement(PIPE_OPERATOR)('|')(11,12)
+          PsiWhiteSpace(' ')(12,13)
+          ElixirMatchedAssociationOperationImpl(MATCHED_ASSOCIATION_OPERATION)(13,25)
+            ElixirVariableImpl(VARIABLE)(13,16)
+              PsiElement(identifier)('key')(13,16)
+            PsiWhiteSpace(' ')(16,17)
+            PsiElement(ASSOCIATION_OPERATOR)('=>')(17,19)
+            PsiWhiteSpace(' ')(19,20)
+            ElixirVariableImpl(VARIABLE)(20,25)
+              PsiElement(identifier)('value')(20,25)
           PsiElement(,)(',')(25,26)
         PsiWhiteSpace(' ')(26,27)
         PsiElement(CLOSING_CURLY)('}')(27,28)
